+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 ### New Features
  - add an option to attach files when schedule a mail action
  - for `qualtricsProject` list on right click open edit or view action in a new tab
+ - for `qualtricsActions` list on right click open edit or view action in a new tab
+ - for `qualtricsSurveys` list on right click open edit or view action in a new tab
 
 # v1.0.2
 ### Bugfix